+++ lib/grell/capybara_driver.rb
@@ -12,8 +12,9 @@ module Grell
     end
 
     def setup_capybara
+      @poltergeist_driver = nil
       Capybara.register_driver :poltergeist_crawler do |app|
-        Capybara::Poltergeist::Driver.new(app, {
+        @poltergeist_driver = Capybara::Poltergeist::Driver.new(app, {
           js_errors: false,
           inspector: false,
           phantomjs_logger: open('/dev/null'),
@@ -28,6 +29,7 @@ module Grell
         "DNT" => 1,
         "User-Agent" => USER_AGENT
       }
+      @poltergeist_driver
     end
   end